pyproject.toml
[tool.poetry]
name = "task-processors"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dramatiq = {extras = ["redis", "watch"], version = "^1.16.0"}
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
redis = "^5.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
